Fix computation of probabilities when end of message reached
Previously if one of the possible continuations of the prefix was "end of message", this was included in the computation of continuation probabilities, even though it was removed from the suggestions in the end. This caused the other 'proper' continuations to have probabilities lower than they should and in extreme cases (when "end of message" was very probable) in showing no suggestions at all even though these existed.

This is now fixed. The "end of message" is not shown among suggestions and the probabilities of proper continuations add up to 100%.